add tests and stories for AlergyIntolerance component
