+++ src/components/resources/AllergyIntolerance/index.js
@@ -1,0 +1,3 @@
+import AllergyIntolerance from './AllergyIntolerance';
+
+export default AllergyIntolerance;
